+++ lib/nsrr/commands/download.rb
@@ -1,6 +1,7 @@
 require 'colorize'
 
 require 'nsrr/models/all'
+require 'nsrr/helpers/authorization'
 
 module Nsrr
   module Commands
@@ -11,7 +12,7 @@ module Nsrr
         end
       end
 
-      attr_reader :dataset_slug, :folder, :file_comparison, :depth
+      attr_reader :token, :dataset_slug, :folder, :file_comparison, :depth
 
       def initialize(argv)
         (@token, argv) = parse_parameter_with_value(argv, ['token'], '')
@@ -21,19 +22,23 @@ module Nsrr
         @folder = (argv[1].to_s.split('/')[1..-1] || []).join('/')
       end
 
+      # Run with Authorization
       def run
-        @dataset = Dataset.find @dataset_slug
-        if @dataset
-          @dataset.download_token = @token
-          @dataset.download(@folder, depth: @depth, method: @file_comparison)
-        else
-          if @dataset_slug == nil
-            puts "Please specify a dataset: " + "nsrr download DATASET".colorize(:white)
-          else
-            puts "The dataset " + "#{@dataset_slug}".colorize(:white) + " was not found."
-          end
+        if @dataset_slug == nil
+          puts "Please specify a dataset: " + "nsrr download DATASET".colorize(:white)
           puts "Read more on the download command here:"
           puts "  " + "https://github.com/nsrr/nsrr-gem".colorize( :blue ).on_white.underline
+        else
+          @token = Nsrr::Helpers::Authorization.get_token(@token) if @token.to_s == ''
+          @dataset = Dataset.find(@dataset_slug, @token)
+          if @dataset
+            @dataset.download(@folder, depth: @depth, method: @file_comparison)
+          else
+            puts "\nThe dataset " + "#{@dataset_slug}".colorize(:white) + " was not found."
+            auth_section = (@token.to_s == '' ? '' : "/a/#{@token}" )
+            datasets = Nsrr::Helpers::JsonRequest.get("#{Nsrr::WEBSITE}#{auth_section}/datasets.json")
+            puts "Did you mean one of: #{datasets.collect{|d| d['slug'].colorize(:white)}.sort.join(', ')}" if datasets and datasets.size > 0
+          end
         end
       end