+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - The `nsrr download` command provides better feedback on the validity of the authorization token provided by the user
 - Use of Ruby 2.2.1 is now recommended
 
+### Bug Fixes
+- Fixed an issue preventing public files from being downloaded when an invalid token was entered
+
 ## 0.1.1 (October 23, 2014)
 
 ### Enhancements