+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - The `nsrr download` command now allows users to download files from private datasets
 - The `nsrr download` command provides better feedback on the validity of the authorization token provided by the user
 - Use of Ruby 2.2.2 is now recommended
+- Added text to inform users that input is hidden while entering token
 
 ### Bug Fixes
 - Fixed an issue preventing public files from being downloaded when an invalid token was entered